music/src/test/javascript/spec/app/entities/ads-playlist-my-suffix/ads-playlist-my-suffix.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { take, map } from 'rxjs/operators';
import { AdsPlaylistMySuffixService } from 'app/entities/ads-playlist-my-suffix/ads-playlist-my-suffix.service';
import { IAdsPlaylistMySuffix, AdsPlaylistMySuffix } from 'app/shared/model/ads-playlist-my-suffix.model';

describe('Service Tests', () => {
  describe('AdsPlaylistMySuffix Service', () => {
    let injector: TestBed;
    let service: AdsPlaylistMySuffixService;
    let httpMock: HttpTestingController;
    let elemDefault: IAdsPlaylistMySuffix;
    let expectedResult: IAdsPlaylistMySuffix | IAdsPlaylistMySuffix[] | boolean | null;
    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(AdsPlaylistMySuffixService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new AdsPlaylistMySuffix(0, 'AAAAAAA', 'AAAAAAA', 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);
        service
          .find(123)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a AdsPlaylistMySuffix', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .create(new AdsPlaylistMySuffix())
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp.body));
        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a AdsPlaylistMySuffix', () => {
        const returnedFromService = Object.assign(
          {
            content: 'BBBBBB',
            image: 'BBBBBB',
            playlistId: 1
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);
        service
          .update(expected)
          .pipe(take(1))
          .subscribe(resp => (expectedResult = resp.body));
        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of AdsPlaylistMySuffix', () => {
        const returnedFromService = Object.assign(
          {
            content: 'BBBBBB',
            image: 'BBBBBB',
            playlistId: 1
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        service
          .query()
          .pipe(
            take(1),
            map(resp => resp.body)
          )
          .subscribe(body => (expectedResult = body));
        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a AdsPlaylistMySuffix', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
